basic design done
